# SMS-Onboarding-Unified/api-specification.yaml
openapi: 3.0.3
info:
  title: SMS Onboarding Portal API
  description: |
    RESTful API for the SMS Onboarding Portal that manages vessel onboarding, equipment documentation,
    critical parts management, and integration with the maintenance portal.
  version: 1.0.0
  contact:
    name: SMS Support Team
    email: support@sms-portal.com
  license:
    name: Proprietary
servers:
  - url: https://api.sms-onboarding.com/v1
    description: Production server
  - url: https://staging-api.sms-onboarding.com/v1
    description: Staging server
  - url: http://localhost:4000/v1
    description: Development server
security:
  - bearerAuth: []
...
paths:
...
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      operationId: refreshToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refreshToken
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/logout:
    post:
      tags:
        - Authentication
      summary: Logout user
      operationId: logout
      responses:
        '200':
          description: Logout successful
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/forgot-password:
    post:
      tags:
        - Authentication
      summary: Request password reset
      operationId: forgotPassword
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Password reset email sent
        '404':
          $ref: '#/components/responses/NotFound'
  /auth/reset-password:
    post:
      tags:
        - Authentication
      summary: Reset password with token
      operationId: resetPassword
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - token
                - newPassword
              properties:
                token:
                  type: string
                newPassword:
                  type: string
                  minLength: 8
      responses:
        '200':
          description: Password reset successful
        '400':
          $ref: '#/components/responses/BadRequest'
  /auth/me:
    get:
      tags:
        - Authentication
      summary: Get current user profile
      operationId: getCurrentUser
      responses:
        '200':
          description: User profile retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
    
    patch:
      tags:
        - Authentication
      summary: Update current user profile
      operationId: updateCurrentUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User profile updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # Company Management
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    
    webhookAuth:
      type: apiKey
      in: header
      name: X-Webhook-Signature
...
  schemas:
...
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        rememberMe:
          type: boolean
          default: false
    UpdateUserRequest:
      type: object
      properties:
        fullName:
          type: string
        avatarUrl:
          type: string
          format: uri
        preferences:
          $ref: '#/components/schemas/UserPreferences'
...
    AuthResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
        token:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        fullName:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
        companyId:
          type: string
          format: uuid
        avatarUrl:
          type: string
          format: uri
        isActive:
          type: boolean
        preferences:
          $ref: '#/components/schemas/UserPreferences'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        lastLoginAt:
          type: string
          format: date-time
    UserPreferences:
      type: object
      properties:
        theme:
          type: string
          enum: [light, dark, system]
        language:
          type: string
        notifications:
          type: object
          properties:
            email:
              type: boolean
            push:
              type: boolean
            inApp:
              type: boolean
        defaultView:
          type: string
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true

    # Enums
    UserRole:
      type: string
      enum:
        - ADMIN
        - TECHNICIAN
        - MANAGER
...
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...